chore: types/auth.ts 삭제 후 types/channel.ts로 옮겨 일관성 유지
