+++ frontend/src/components/common/Navbar/NavbarUserInfo.tsx
@@ -1,4 +1,4 @@
-import type { User } from '../../../types/auth'
+import type { User } from '../../../types/channel'
 
 interface NavbarUserInfoProps {
     user: User
