+++ frontend/src/stores/authStore.ts
@@ -1,6 +1,6 @@
 import { create } from 'zustand'
 import { devtools, persist } from 'zustand/middleware'
-import type { User } from '../types/auth'
+import type { User } from '../types/channel'
 
 interface AuthActions {
     setUser: (user: User) => void
