--- frontend/src/types/auth.ts
@@ -1,15 +1,0 @@
-import type { CommonResponse } from './common'
-
-export interface User {
-    memberId: number
-    channelId?: number
-    nickname: string
-    googleEmail: string
-    profileImage: string
-    instagramLink: string | null
-    tiktokLink: string | null
-    facebookLink: string | null
-    twitterLink: string | null
-}
-
-export type ResponseMyProfile = CommonResponse<User>
