+++ frontend/src/types/channel.ts
@@ -1,3 +1,5 @@
+import type { CommonResponse } from './common'
+
 export interface ChannelTargetDto {
     channelId: number
     target: string
@@ -7,3 +9,17 @@ export interface ChannelConceptDto {
     channelId: number
     concept: string
 }
+
+export interface User {
+    memberId: number
+    channelId?: number
+    nickname: string
+    googleEmail: string
+    profileImage: string
+    instagramLink: string | null
+    tiktokLink: string | null
+    facebookLink: string | null
+    twitterLink: string | null
+}
+
+export type ResponseMyProfile = CommonResponse<User>
